refactor: crear package de star wars
