+++ app/src/main/res/layout/activity_star_wars.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StarWarsActivity">
+    tools:context=".starwars.StarWarsActivity">
 
     <Button
         android:id="@+id/fetch_films"
